I pulled the ability to make multiple timers
